wip: udp refactoring
